requirements.txt
pydantic>=2.0
requests>=2.31
PyYAML>=6.0
serpapi
aiohttp>=3.8
firebase-admin>=6.4.0
